Changed some int's into long's.

+++ OBSClient/Messages/MediaInputCursorOffsetRequest.cs
@@ -8,9 +8,9 @@
         public string InputName { get; set; }
 
         [JsonPropertyName("mediaCursorOffset")]
-        public int MediaCursorOffset { get; set; }
+        public long MediaCursorOffset { get; set; }
 
-        public MediaInputCursorOffsetRequest(string inputName, int mediaCursorOffset)
+        public MediaInputCursorOffsetRequest(string inputName, long mediaCursorOffset)
         {
             this.InputName = inputName;
             this.MediaCursorOffset = mediaCursorOffset;
